+++ packages/themes/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
-  entry: ['src/index.ts', 'src/schemas.ts'],
+  entry: ['src/index.ts', 'src/schemas.ts', 'src/color-mode-script.ts'],
   format: ['cjs', 'esm'],
   dts: true,
   target: 'es2022',